AuctionSite
An auction site created with asp.net core 2.0.

The assignment was to create the site using an external (pre-made) API for storing of auction-data.
Then we were asked to create login with asp.net Identity, using different roles. And show views according to
the role, so  in this case, the admin role can view more information than the member role.

#Features
The Nackowski-App has member and admin roles, admin can create auctions and view sales statistics created with Charts.js.
A member can view ongoing actions and filter the acctions according to date/name.

It contains a test project with simple unittests. 
In this project I also tried out cashing.